# src/application/server/docs/swagger.yaml
definitions:
  handler.PostSubmissionBody:
    type: object
...
info:
  contact: {}
paths:
...
  /problem/{slug}/check:
    get:
      consumes:
      - application/json
      description: Check problem slug
      parameters:
      - description: problem slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: Check problem slug
      tags:
      - problem
  /problem/{slug}/package:
    put:
      consumes:
      - application/json
      description: Put problem package
      parameters:
      - description: problem slug
        in: path
        name: slug
        required: true
        type: string
      responses: {}
      summary: Put problem package
      tags:
      - problem
  /problem/{slug}/submission:
    post:
      consumes:
      - application/json
      description: Post submission
      parameters:
      - description: problem slug
        in: path
        name: slug
        required: true
        type: string
      - description: judge request
        in: body
        name: judgeRequest
        required: true
        schema:
          $ref: '#/definitions/handler.PostSubmissionBody'
      responses: {}
      summary: Post submission
      tags:
      - problem
...
swagger: "2.0"